# pages/api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Sharpstudy API",
    "description": "Description of your API",
    "version": "1.0.0",
    "contact": {
      "name": "Sharpstudy",
      "email": "hello@sharpstudy.io"
    }
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "paths": {
...
    "/api/quizzes/instructor/quiz": {
      "post": {
        "summary": "Create a new quiz",
        "description": "Creates a new quiz.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "Title of the quiz"
                  },
                  "description": {
                    "type": "string",
                    "description": "Description of the quiz"
                  },
                  "courseId": {
                    "type": "string",
                    "description": "ID of the course associated with the quiz"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Quiz created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "quiz": {
                      "type": "object",
                      "$ref": "#/components/schemas/Quiz"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error_code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Quizzes"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "summary": "Delete a Quiz and related data",
        "description": "Delete a Quiz and related data",
        "parameters": [
          {
            "in": "query",
            "name": "quizId",
            "required": true,
            "description": "ID of the quiz to be deleted.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error_code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "Quizzes"
        ]
      },
      "get": {
        "summary": "Find one quiz",
        "description": "Retrieve quiz for the authenticated user.",
        "tags": [
          "Quizzes"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "quizId",
            "required": true,
            "description": "ID of the Quiz to retrieve.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "quizzes": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error_code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update an existing quiz",
        "description": "Endpoint to update an existing quiz.",
        "tags": [
          "Quizzes"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "quizId",
            "required": true,
            "description": "ID of the Quiz to update.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "Title of the quiz to update"
                  },
                  "description": {
                    "type": "string",
                    "description": "Description of the quiz to update"
                  }
                },
                "required": [
                  "courseId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error_code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Quiz": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "courseId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "title",
          "description",
          "courseId"
        ]
      },
...
    }
  }
}